requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
scikit-learn
tqdm
Pillow
torchinfo
shutil
